+++ NEWS.md
@@ -5,9 +5,11 @@
 * belamkar.augmented (augmented design at multiple locations)
 * evans.sugarcane.uniformity
 * goulden.barley.uniformity (expanded from the previously published 20x20 to 48x48)
+* hansen.multi.uniformity (multi-year uniformity trials)
 * kirk.potato (20 varieties with 15 reps)
 * smith.wheat.uniformity (classic data for measuring field heterogeneity)
 * vishnaadevi.rice.uniformity
+
 
 # agridat 1.21 - Jun 2022